Refactor type_definition method
Converts the old type_definition method from the old libbpf-cargo wrapper to the
new wrapper. I delete the old method in order to generate a diff where
both can be compared.
