+++ libbpf-cargo/src/btf/btf.rs
@@ -546,337 +546,6 @@ impl Btf {
         })
     }
 
-    /// Returns rust type definition of `ty` in string format, including dependent types.
-    ///
-    /// `ty` must be a struct, union, enum, or datasec type.
-    pub fn type_definition(&self, type_id: u32) -> Result<String> {
-        let next_type = |mut id| -> Result<Option<u32>> {
-            loop {
-                match self.type_by_id(id)? {
-                    BtfType::Struct(_)
-                    | BtfType::Union(_)
-                    | BtfType::Enum(_)
-                    | BtfType::Datasec(_) => return Ok(Some(id)),
-                    BtfType::Ptr(t) => id = t.pointee_type,
-                    BtfType::Array(t) => id = t.val_type_id,
-                    BtfType::Volatile(t) => id = t.type_id,
-                    BtfType::Const(t) => id = t.type_id,
-                    BtfType::Restrict(t) => id = t.type_id,
-                    BtfType::Typedef(t) => id = t.type_id,
-                    BtfType::DeclTag(t) => id = t.type_id,
-                    BtfType::TypeTag(t) => id = t.type_id,
-                    BtfType::Void
-                    | BtfType::Int(_)
-                    | BtfType::Float(_)
-                    | BtfType::Fwd(_)
-                    | BtfType::Func(_)
-                    | BtfType::FuncProto(_)
-                    | BtfType::Var(_) => return Ok(None),
-                }
-            }
-        };
-
-        let is_terminal = |id| -> Result<bool> {
-            match self.type_by_id(id)?.kind() {
-                BtfKind::Struct | BtfKind::Union | BtfKind::Enum | BtfKind::Datasec => Ok(false),
-                BtfKind::Void
-                | BtfKind::Int
-                | BtfKind::Float
-                | BtfKind::Ptr
-                | BtfKind::Array
-                | BtfKind::Fwd
-                | BtfKind::Typedef
-                | BtfKind::Volatile
-                | BtfKind::Const
-                | BtfKind::Restrict
-                | BtfKind::Func
-                | BtfKind::FuncProto
-                | BtfKind::Var
-                | BtfKind::DeclTag
-                | BtfKind::TypeTag => Ok(true),
-            }
-        };
-
-        ensure!(
-            !is_terminal(type_id)?,
-            "Tried to print type definition for terminal type"
-        );
-
-        // Process dependent types until there are none left.
-        //
-        // When we hit a terminal, we write out some stuff. A non-terminal adds more types to
-        // the queue.
-        let mut def = String::new();
-        let mut dependent_types = vec![type_id];
-        let mut processed = BTreeSet::new();
-        while !dependent_types.is_empty() {
-            let type_id = dependent_types.remove(0);
-            if processed.contains(&type_id) {
-                continue;
-            } else {
-                processed.insert(type_id);
-            }
-
-            let ty = self.type_by_id(type_id)?;
-
-            match ty {
-                BtfType::Struct(t) | BtfType::Union(t) => {
-                    let packed = self.is_struct_packed(type_id, t)?;
-
-                    // fields in the aggregate
-                    let mut agg_content: Vec<String> = Vec::new();
-
-                    // structs with arrays > 32 length need to impl Default
-                    // rather than #[derive(Default)]
-                    let mut impl_default: Vec<String> = Vec::new(); // output for impl Default
-                    let mut gen_impl_default = false; // whether to output impl Default or use #[derive]
-
-                    let mut offset = 0; // In bytes
-                    for member in &t.members {
-                        ensure!(
-                            member.bit_size == 0 && member.bit_offset % 8 == 0,
-                            "Struct bitfields not supported"
-                        );
-
-                        let field_ty_id = self.skip_mods_and_typedefs(member.type_id)?;
-                        if let Some(next_ty_id) = next_type(field_ty_id)? {
-                            dependent_types.push(next_ty_id);
-                        }
-
-                        // Add padding as necessary
-                        if t.is_struct {
-                            let padding = self.required_padding(
-                                offset,
-                                member.bit_offset as usize / 8,
-                                member.type_id,
-                                packed,
-                            )?;
-
-                            if padding != 0 {
-                                agg_content
-                                    .push(format!(r#"    __pad_{offset}: [u8; {padding}],"#,));
-
-                                impl_default.push(format!(
-                                    r#"            __pad_{offset}: [u8::default(); {padding}]"#,
-                                ));
-                            }
-
-                            if let BtfType::Array(ft) = self.type_by_id(field_ty_id)? {
-                                if ft.nelems > 32 {
-                                    gen_impl_default = true
-                                }
-                            }
-                        }
-
-                        match self.type_default(field_ty_id) {
-                            Ok(def) => {
-                                impl_default.push(format!(
-                                    r#"            {field_name}: {field_ty_str}"#,
-                                    field_name = member.name,
-                                    field_ty_str = def
-                                ));
-                            }
-                            Err(e) => {
-                                if gen_impl_default || !t.is_struct {
-                                    bail!("Could not construct a necessary Default Impl: {}", e);
-                                }
-                            }
-                        };
-
-                        // Set `offset` to end of current var
-                        offset = ((member.bit_offset / 8) + self.size_of(field_ty_id)?) as usize;
-
-                        let field_ty_str = self.type_declaration(field_ty_id)?;
-                        let field_name = if !member.name.is_empty() {
-                            member.name.to_string()
-                        } else {
-                            field_ty_str.clone()
-                        };
-
-                        agg_content.push(format!(r#"    pub {field_name}: {field_ty_str},"#));
-                    }
-
-                    if t.is_struct {
-                        let struct_size = t.size as usize;
-                        let padding =
-                            self.required_padding(offset, struct_size, type_id, packed)?;
-                        if padding != 0 {
-                            agg_content.push(format!(r#"    __pad_{offset}: [u8; {padding}],"#,));
-                            impl_default.push(format!(
-                                r#"            __pad_{offset}: [u8::default(); {padding}]"#,
-                            ));
-                        }
-                    }
-
-                    if !gen_impl_default && t.is_struct {
-                        writeln!(def, r#"#[derive(Debug, Default, Copy, Clone)]"#)?;
-                    } else if t.is_struct {
-                        writeln!(def, r#"#[derive(Debug, Copy, Clone)]"#)?;
-                    } else {
-                        writeln!(def, r#"#[derive(Copy, Clone)]"#)?;
-                    }
-
-                    let aggregate_type = if t.is_struct { "struct" } else { "union" };
-                    let packed_repr = if packed { ", packed" } else { "" };
-
-                    writeln!(def, r#"#[repr(C{packed_repr})]"#)?;
-                    writeln!(
-                        def,
-                        r#"pub {agg_type} {name} {{"#,
-                        agg_type = aggregate_type,
-                        name = t.name,
-                    )?;
-
-                    for field in agg_content {
-                        writeln!(def, "{field}")?;
-                    }
-                    writeln!(def, "}}")?;
-
-                    // if required write a Default implementation for this struct
-                    if gen_impl_default {
-                        writeln!(def, r#"impl Default for {} {{"#, t.name)?;
-                        writeln!(def, r#"    fn default() -> Self {{"#)?;
-                        writeln!(def, r#"        {} {{"#, t.name)?;
-                        for impl_def in impl_default {
-                            writeln!(def, r#"{impl_def},"#)?;
-                        }
-                        writeln!(def, r#"        }}"#)?;
-                        writeln!(def, r#"    }}"#)?;
-                        writeln!(def, r#"}}"#)?;
-                    } else if !t.is_struct {
-                        // write a Debug implementation for a union
-                        writeln!(def, r#"impl std::fmt::Debug for {} {{"#, t.name)?;
-                        writeln!(
-                            def,
-                            r#"    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {{"#
-                        )?;
-                        writeln!(def, r#"        write!(f, "(???)")"#)?;
-                        writeln!(def, r#"    }}"#)?;
-                        writeln!(def, r#"}}"#)?;
-
-                        // write a Default implementation for a union
-                        writeln!(def, r#"impl Default for {} {{"#, t.name)?;
-                        writeln!(def, r#"    fn default() -> Self {{"#)?;
-                        writeln!(def, r#"        {} {{"#, t.name)?;
-                        writeln!(def, r#"{},"#, impl_default[0])?;
-                        writeln!(def, r#"        }}"#)?;
-                        writeln!(def, r#"    }}"#)?;
-                        writeln!(def, r#"}}"#)?;
-                    }
-                }
-                BtfType::Enum(t) => {
-                    let repr_size = match t.size {
-                        1 => "8",
-                        2 => "16",
-                        4 => "32",
-                        8 => "64",
-                        16 => "128",
-                        _ => bail!("Invalid enum size: {}", t.size),
-                    };
-
-                    let mut signed = "u";
-                    for value in &t.values {
-                        if value.value < 0 {
-                            signed = "i";
-                            break;
-                        }
-                    }
-
-                    writeln!(
-                        def,
-                        r#"#[derive(Debug, Copy, Clone, Default, PartialEq, Eq)]"#
-                    )?;
-                    writeln!(def, r#"#[repr({signed}{repr_size})]"#)?;
-                    writeln!(def, r#"pub enum {name} {{"#, name = t.name)?;
-
-                    for (i, value) in t.values.iter().enumerate() {
-                        if i == 0 {
-                            writeln!(def, r#"    #[default]"#)?;
-                        }
-
-                        writeln!(
-                            def,
-                            r#"    {name} = {value},"#,
-                            name = value.name,
-                            value = value.value,
-                        )?;
-                    }
-
-                    writeln!(def, "}}")?;
-                }
-                BtfType::Datasec(t) => {
-                    let mut sec_name = t.name.to_string();
-                    if sec_name.is_empty() || !sec_name.starts_with('.') {
-                        bail!("Datasec name is invalid: {}", sec_name);
-                    }
-                    sec_name.remove(0);
-
-                    writeln!(def, r#"#[derive(Debug, Copy, Clone)]"#)?;
-                    writeln!(def, r#"#[repr(C)]"#)?;
-                    writeln!(def, r#"pub struct {sec_name} {{"#)?;
-
-                    let mut offset: u32 = 0;
-                    for datasec_var in &t.vars {
-                        let var = match self.type_by_id(datasec_var.type_id)? {
-                            BtfType::Var(v) => {
-                                if v.linkage == BtfVarLinkage::Static {
-                                    // do not output Static Var
-                                    continue;
-                                }
-
-                                if let Some(next_ty_id) = next_type(v.type_id)? {
-                                    dependent_types.push(next_ty_id);
-                                }
-
-                                v
-                            }
-                            _ => bail!("BTF is invalid! Datasec var does not point to a var"),
-                        };
-
-                        let padding = self.required_padding(
-                            offset as usize,
-                            datasec_var.offset as usize,
-                            var.type_id,
-                            false,
-                        )?;
-                        if padding != 0 {
-                            writeln!(def, r#"    __pad_{offset}: [u8; {padding}],"#)?;
-                        }
-
-                        // Set `offset` to end of current var
-                        offset = datasec_var.offset + datasec_var.size;
-
-                        writeln!(
-                            def,
-                            r#"    pub {var_name}: {var_type},"#,
-                            var_name = var.name,
-                            var_type = self.type_declaration(var.type_id)?
-                        )?;
-                    }
-
-                    writeln!(def, "}}")?;
-                }
-                BtfType::Void
-                | BtfType::Ptr(_)
-                | BtfType::Func(_)
-                | BtfType::Int(_)
-                | BtfType::Float(_)
-                | BtfType::Array(_)
-                | BtfType::Fwd(_)
-                | BtfType::Typedef(_)
-                | BtfType::FuncProto(_)
-                | BtfType::Var(_)
-                | BtfType::Volatile(_)
-                | BtfType::Const(_)
-                | BtfType::Restrict(_)
-                | BtfType::DeclTag(_)
-                | BtfType::TypeTag(_) => bail!("Invalid type: {}", ty),
-            }
-        }
-
-        Ok(def)
-    }
-
     pub fn skip_mods_and_typedefs(&self, mut type_id: u32) -> Result<u32> {
         loop {
             match self.type_by_id(type_id)? {
